Modify encoding scheme
Add a start sequence token

In particular, added the "!" character to the one-hot encoding
to indicate a start token. Modified the encoder/decoder to account
for this and did some slight refactoring. Most of the changes
were because the "!" character maps to 0, which is where "A" used
to map so tests had to be fixed