Modify data processing pipeline extraction and encoding
We have broken down the SlidingWindowExtractor to just extract
kmer strings. We then encode the kmer strings into a matrix
of integers and finally do a one-hot encoding.

This was done because I'd like to use the output of the new
SlidingWindowExtractor to get some statistics on input-output
mappings and it would have been a bit harder if the input and
outputs were matrices of characters rather than just arrays of
strings.

One hot encoding can now proceed through a simple nested array
access

The bottleneck now is the integer encoding (when the fill_in_the_blanks
option is ticked) and somewhat the one-hot encoding if we need to
add some "padding" to the input. In the best case when neither parameter
is ticked and the runtime is actually better than before.

In the worst case when both parameters are ticked, the runtime
should be no worse than before