Implement more extensive training set
We now validate over all data points used in training for an
epoch at the end of the epoch with the final model weights at
that epoch. It should give a less biased metric of training error
than just the training error for the final batch (which we just
trained on). However, this adds a lot of overhead to the training
process. Might need to timestamp how long it takes to do the prediction
to see if it's because of how we're slicing lists or if it's
actually just the model prediction step.